반려동물, 회원에 관한 작업 끝, 진료기록 작성중(큰 오류 나서 대기)

+++ 02project_0722/src/main/webapp/WEB-INF/views/member/memberList.jsp
@@ -70,7 +70,7 @@
                             <td class="py-2 text-center">
                                 <a href="${pageContext.request.contextPath}/member/modify.do?m_idx=${member.m_idx}" class="text-blue-500 hover:underline">수정</a>
                             </td>
-                            <td class="py-2 text-center">
+                            <td class="py-2 text-center" onclick="return confirm('정말 삭제하시겠습니까?');">
                                 <a href="${pageContext.request.contextPath}/member/delete.do?m_idx=${member.m_idx}" class="text-red-500 hover:underline">삭제</a>
                             </td>                    
                         </tr>
